+++ CHANGELOG.md
@@ -133,3 +133,4 @@
 * Define `yoklKL` operator
 * Define `kl` operator
 * Define `Empty` patter synonym
+* Define `Layered` type family